# docs/api/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Wizerproperties
paths:
...
  /building/api/list/popular/:
    get:
      tags:
        - Building
      summary: "Popular buildings list"
      description: "**User type**: both"
      operationId: Popular
      produces:
        - "application/json"
      parameters:
        - name: "page_size"
          in: "query"
          type: "number"
        - name: "page"
          in: "query"
          type: "number"
        - name: "towards"
          description: "Its control page (e.g 'search' then show fields only for search page)"
          in: "query"
          type: "string"
        - name: "lowest_price"
          description: "Filtered by"
          in: "query"
          type: "number"
        - name: "highest_price"
          description: "Filtered by"
          in: "query"
          type: "number"
        - name: "type"
          description: "Filtered by and choice will be 'condos' or 'house'."
          in: "query"
          type: "string"
        - name: "quota"
          description: "Choice will be 'thai' or 'foreign'."
          in: "query"
          type: "string"
        - name: "furnishing"
          description: "Choice will be 'partly' or 'fully'."
          in: "query"
          type: "string"
        - name: "have_freehold"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_leasehold"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_infinity_pool"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_pets_allowed"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_guard_house"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_sauna"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_sky_lounge"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_grocery"
          description: "Filtered by"
          in: "query"
          type: "boolean"
        - name: "have_fitness_area"
          description: "Filtered by"
          in: "query"
          type: "boolean"
      responses:
        200:
          description: Popular Buildings list
          schema:
            $ref: "#/definitions/BuildingVariousFeatureListResponse"
        405:
          description: Method not allowed(Method ain't supported by this URL)
          schema:
            $ref: "#/definitions/commonSingleErrorResponse"
        500:
          description: Internal Server Error(Something has gone wrong on the website's server)
        501:
          description: Not Implemented(Web server does not support the functionality required for access and does not recognize the request method.)
        502:
          description: Bad Gateway(Received an invalid response from the upstream server)
        503:
          description: Service Unavailable(Server is down for maintenance or overloaded)
        504:
          description: Taking more time to get a response
...
  /property/api/list/popular/:
    get:
      tags:
        - Property
      summary: "Popular properties list"
      description: "**User type**: both"
      operationId: popular
      produces:
        - "application/json"
      parameters:
        - name: "page_size"
          in: "query"
          type: "number"
        - name: "page"
          in: "query"
          type: "number"
        - name: "towards"
          description: "Its control page (e.g 'search' then show fields only for search page)"
          in: "query"
          type: "string"
      responses:
        200:
          description: Properties list
          schema:
            $ref: "#/definitions/PropertyVariousFeatureListResponse"
        405:
          description: Method not allowed(Method ain't supported by this URL)
          schema:
            $ref: "#/definitions/commonSingleErrorResponse"
        500:
          description: Internal Server Error(Something has gone wrong on the website's server)
        501:
          description: Not Implemented(Web server does not support the functionality required for access and does not recognize the request method.)
        502:
          description: Bad Gateway(Received an invalid response from the upstream server)
        503:
          description: Service Unavailable(Server is down for maintenance or overloaded)
        504:
          description: Taking more time to get a response
...
definitions:
  commonSingleErrorResponse:
    type: "object"
    properties:
      detail:
        type: "string"
...
  BuildingVariousFeatureListResponse:
    type: object
    properties:
      count:
        type: "number"
      page_size:
        type: "number"
      next:
        type: "number"
      previous:
        type: "number"
      total_page:
        type: "number"
      results:
        type: array
        items:
          type: object
          properties:
            id:
              type: "number"
            title:
              type: "string"
            description:
              type: "string"
            type:
              type: "string"
            total_units_for_sale:
              type: "number"
            address:
              type: "string"
            project_total_area:
              type: "number"
            total_floors:
              type: "number"
            facility_view:
              type: "string"
            location_view:
              type: "string"
            latitude:
              type: "number"
            longitude:
              type: "number"
            default_image:
              type: "string"
...
  PropertyVariousFeatureListResponse:
    type: object
    properties:
      count:
        type: "number"
      page_size:
        type: "number"
      next:
        type: "number"
      previous:
        type: "number"
      total_page:
        type: "number"
      results:
        type: array
        items:
          type: object
          properties:
            id:
              type: "number"
            title:
              type: "string"
            description:
              type: "string"
            price:
              type: "number"
            price_per_sqm:
              type: "number"
            floor_number:
              type: "string"
            unit_area:
              type: "number"
            number_of_bedroom:
              type: "number"
            number_of_bathroom:
              type: "number"
            building_type:
              type: "string"
            building_address:
              type: "string"
            is_compared:
              type: "boolean"
            is_favorited:
              type: "boolean"
            default_image:
              type: "string"
            discount_period:
              type: "string"
...
host: localhost:8000
schemes:
  - https
  - http